fix correct langfuse/ folder structure
